MNIST_dl/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli==1.27.38
flake8
python-dotenv>=0.5.1

absl-py==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
hydra-core==1.3.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy
oauthlib==3.2.2
omegaconf==2.3.0
packaging==21.3
pandas
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==1.3.1
snakeviz==2.1.1
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.0
tinycss2==1.2.1
torch==1.13.1
torch-tb-profiler==0.4.1
torchvision==0.14.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.7


